Add semantics
Make the line between opponent and player score more visible and add labels for the two parts.
Add semantics to HTML (e.g. list elements in a <ul> tag, <span> tags). According to the front-enders I have learned from, this is important and Google likes this very much.
Added spaces and adjusted indentation - it was an automatic reaction and I saw that in git later on. 😄